# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QRX API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get doctor profile (public)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "404":
          description: Not found
    patch:
      operationId: updateDoctor
      tags: [doctors]
      summary: Update doctor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves object entities uploaded via presigned URLs. These can optionally
        be protected with authentication or ACL checks based on the use case.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ---- QUEUE DISPLAY DEVICES ----
...
components:
  schemas:
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
        category:
          type: string
          enum: [doctors, prescriptions, chat, banners, blog, shop, general]
          description: |
            Upload category — determines the storage subdirectory. Defaults to
            "general" if omitted. Use this to route files to the correct folder:
            doctors (profile photos), prescriptions (PDFs/lab reports),
            chat (message attachments), banners (hero/banner images),
            blog (post cover images), shop (product images).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          description: Upload destination URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/doctors/uuid`). Store this in your database.
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Doctor:
      type: object
      required: [id, name, email, approvalStatus]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: string
        photoUrl:
          type: ["string", "null"]
        degree:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        specialtyId:
          type: ["integer", "null"]
        specialtyName:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        experience:
          type: ["integer", "null"]
        chamberAddress:
          type: ["string", "null"]
        visitingTime:
          type: ["string", "null"]
        chamberAddress2:
          type: ["string", "null"]
        visitingTime2:
          type: ["string", "null"]
        consultationFee:
          type: ["integer", "null"]
        bmdcNumber:
          type: ["string", "null"]
        bmdcValidityYears:
          type: ["integer", "null"]
        subscriptionFee:
          type: ["integer", "null"]
        currency:
          type: string
          enum: [BDT, USD]
        approvalStatus:
          type: string
          enum: [pending, approved, rejected, expired]
        isFeatured:
          type: boolean
        isSenior:
          type: boolean
        about:
          type: ["string", "null"]
        services:
          type: ["string", "null"]
        education:
          type: ["string", "null"]
        onlineStatus:
          type: ["string", "null"]
          enum: [online, offline, busy, vacation, null]
        isVerified:
          type: boolean
        onlineConsultationAvailable:
          type: boolean
        emergencyAvailable:
          type: boolean
        countryId:
          type: ["integer", "null"]
        countryName:
          type: ["string", "null"]
        cityId:
          type: ["integer", "null"]
        cityName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DoctorUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
        degree:
          type: string
        departmentId:
          type: integer
        specialtyId:
          type: integer
        locationId:
          type: integer
        experience:
          type: integer
        chamberAddress:
          type: string
        visitingTime:
          type: string
        chamberAddress2:
          type: string
        visitingTime2:
          type: string
        consultationFee:
          type: integer
        countryId:
          type: integer
        cityId:
          type: integer
        about:
          type: string
        services:
          type: string
        education:
          type: string
        isFeatured:
          type: boolean
        isSenior:
          type: boolean
        approvalStatus:
          type: string
...